replace formatting.ts and formattingHelpersv2.ts
