+++ packages/language-support/src/tests/formatting/cursorpos.test.ts
@@ -4,7 +4,7 @@
  * since it would be difficult to consolidate the new and the old version
  */
 
-import { formatQuery } from '../../formatting/formattingv2';
+import { formatQuery } from '../../formatting/formatting';
 
 // The @ represents the position of the cursor
 describe('tests for correct cursor position', () => {
